Move objects to dedicated files
